3d plot now working with cut surface, but only for dense data. TODO: support positions also.
