+++ python/src/scipp/plot/controller3d.py
@@ -15,7 +15,7 @@ class PlotController3d(PlotController):
     def __init__(self,
                  *args,
                  pixel_size=None,
-                 # positions=None,
+                 positions=None,
                  **kwargs):
 
         super().__init__(*args,
@@ -39,6 +39,16 @@ class PlotController3d(PlotController):
                     "label":
                     name_with_unit(coord, name=xyz.upper())
                 }
+
+    def initialise_model(self):
+        self.model.initialise(self.panel.get_cut_options())
+
+    def connect_panel(self):
+        self.panel.connect({
+            "update_opacity": self.update_opacity,
+            "update_depth_test": self.update_depth_test,
+            "update_cut_surface": self.update_cut_surface
+            })
 
     def _get_axes_parameters(self):
         axparams = {}
@@ -83,7 +93,7 @@ class PlotController3d(PlotController):
     def rescale_to_data(self, button=None):
         super().rescale_to_data()
         new_values = self.model.get_slice_values(
-            mask_info=self._get_mask_info())
+            mask_info=self.get_masks_info())
         self.view.update_data(new_values)
 
     def toggle_mask(self, change=None):
@@ -91,5 +101,5 @@ class PlotController3d(PlotController):
         Show/hide masks
         """
         new_values = self.model.get_slice_values(
-            mask_info=self._get_mask_info())
+            mask_info=self.get_masks_info())
         self.view.update_data(new_values)
