+++ python/src/scipp/plot/panel1d.py
@@ -3,32 +3,34 @@
 # @author Neil Vaytet
 
 # Scipp imports
+from .panel import PlotPanel
 from .._utils import make_random_color
 
 # Other imports
 import ipywidgets as ipw
 
 
-class PlotPanel1d:
+class PlotPanel1d(PlotPanel):
     def __init__(self, data_names):
+        super().__init__()
 
         # self.controller = controller
-        self.container = ipw.VBox()
+        # self.container = ipw.VBox()
         self.keep_buttons = {}
         self.data_names = data_names
         self.slice_label = None
         self.counter = -1
         self.interface = {}
 
-    def _ipython_display_(self):
-        return self._to_widget()._ipython_display_()
+    # def _ipython_display_(self):
+    #     return self._to_widget()._ipython_display_()
 
-    def _to_widget(self):
-        return self.container
+    # def _to_widget(self):
+    #     return self.container
 
-    def connect(self, callbacks):
-        for key, func in callbacks.items():
-            self.interface[key] = func
+    # def connect(self, callbacks):
+    #     for key, func in callbacks.items():
+    #         self.interface[key] = func
 
 
     def make_keep_button(self):
